+++ L2J_Server_BETA/java/com/l2jserver/gameserver/model/actor/stat/ControllableAirShipStat.java
@@ -34,7 +34,7 @@ public class ControllableAirShipStat extends VehicleStat
 	}
 	
 	@Override
-	public float getMoveSpeed()
+	public double getMoveSpeed()
 	{
 		if (getActiveChar().isInDock() || (getActiveChar().getFuel() > 0))
 		{
